requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
ecdsa==0.17.0
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
Mako==1.2.0
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.20.0
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
sendgrid==6.9.7
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starkbank-ecdsa==2.0.3
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
wincertstore==0.2
zipp==3.8.0
